+++ parser/src/sema/import_resolver.rs
@@ -47,25 +47,28 @@ impl VisitMut<SemanticAnalysisError> for ImportResolver<'_> {
         for import in imports.values_mut() {
             match import {
                 Import::All { module: from } => {
-                    let imported_from = match self
-                        .library
-                        .get(from)
-                        .ok_or(SemanticAnalysisError::ImportUndefined(*from))
-                    {
-                        Ok(value) => value,
-                        Err(err) => return ControlFlow::Break(err),
-                    };
-                    for export in imported_from.exports() {
-                        let name = export.name();
-                        let item = Identifier::new(from.span(), name.name());
-                        self.import(module, *from, item, export)?;
+                    let submodules = self.library.get_submodules_of(from);
+                    for submodule in submodules {
+                        let imported_from = match self
+                            .library
+                            .get(&submodule)
+                            .ok_or(SemanticAnalysisError::ImportUndefined(submodule.clone()))
+                        {
+                            Ok(value) => value,
+                            Err(err) => return ControlFlow::Break(err),
+                        };
+                        for export in imported_from.exports() {
+                            let name = export.name();
+                            let item = Identifier::new(from.span(), name.name());
+                            self.import(module, from.clone(), item, export)?;
+                        }
                     }
                 },
                 Import::Partial { module: from, items } => {
                     let imported_from = match self
                         .library
                         .get(from)
-                        .ok_or(SemanticAnalysisError::ImportUndefined(*from))
+                        .ok_or(SemanticAnalysisError::ImportUndefined(from.clone()))
                     {
                         Ok(value) => value,
                         Err(err) => return ControlFlow::Break(err),
@@ -77,7 +80,7 @@ impl VisitMut<SemanticAnalysisError> for ImportResolver<'_> {
                         // with the item in the set, not the span associated with the
                         // export.
                         if let Some(item) = items.get(&name) {
-                            self.import(module, *from, *item, export)?;
+                            self.import(module, from.clone(), *item, export)?;
                         }
                     }
                 },